move chat card to component
