+++ src/components/playQuiz/Lobby.tsx
@@ -1,14 +1,13 @@
 import Background from "@ui/Background";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
-import * as Tabs from "@radix-ui/react-tabs";
 import { useGameStore } from "@utils/zustand/gameStore";
 import { socket } from "@utils/websocket/socket";
-import Chat from "./Chat";
 import TeamCard from "./TeamCard";
 import NavBar from "./NavBar";
 import JoinForm from "./JoinForm";
 import InformationCard from "./InformationCard";
+import ChatCard from "./ChatCard";
 
 export interface IPlayer {
   id: string;
@@ -44,10 +43,6 @@ const Lobby = () => {
     name?: string;
     team: string | null;
   }>();
-  const [messages, setMessages] = useState<IMessage[]>([]);
-  const [teamMessages, setTeamMessages] = useState<IMessage[]>([]);
-  const [numOfUnreadMessages, setNumOfUnreadMessages] = useState(0);
-  const [currentTab, setCurrentTab] = useState("global");
   const [wasKicked, setWasKicked] = useState(false);
 
   const userName = useGameStore((state) => state.userName);
@@ -55,6 +50,8 @@ const Lobby = () => {
   const quizName = useGameStore((state) => state.quizName);
   const numOfQuestions = useGameStore((state) => state.numOfQuestions);
   const numOfTeams = useGameStore((state) => state.numOfTeams);
+  const messages = useGameStore((state) => state.messages);
+  const setMessages = useGameStore((state) => state.setMessages);
   const isQuizMaster = !!(userName && socketId);
 
   const colors = ["#991b1b", "#1e40af", "#166534", "#6b21a8", "#854d0e"];
@@ -85,7 +82,7 @@ const Lobby = () => {
 
     const onLobbyJoin = (data: ILobby, message: IMessage) => {
       setLobby(data);
-      setMessages((oldMessages) => [...oldMessages, message]);
+      setMessages([...messages, message]);
     };
 
     const onTeamJoin = (data: ILobby) => {
@@ -94,20 +91,6 @@ const Lobby = () => {
 
     const onTeamNameEdited = (data: ILobby) => {
       setLobby(data);
-    };
-
-    const onGlobalMessage = (message: IMessage) => {
-      setMessages((oldMessages) => [...oldMessages, message]);
-      if (currentTab !== "global") {
-        setNumOfUnreadMessages(numOfUnreadMessages + 1);
-      }
-    };
-
-    const onTeamMessage = (message: IMessage) => {
-      setTeamMessages((oldMessages) => [...oldMessages, message]);
-      if (currentTab !== "team") {
-        setNumOfUnreadMessages(numOfUnreadMessages + 1);
-      }
     };
 
     const onPlayerDisconnect = (data: { lobby: ILobby; socketId: string }) => {
@@ -138,8 +121,6 @@ const Lobby = () => {
 
     socket.on("joined-lobby", onLobbyJoin);
     socket.on("joined-team", onTeamJoin);
-    socket.on("global-message-received", onGlobalMessage);
-    socket.on("team-message-received", onTeamMessage);
     socket.on("team-name-edited", onTeamNameEdited);
     socket.on("player-disconnect", onPlayerDisconnect);
     socket.on("teams-reset", onTeamsReset);
@@ -148,14 +129,12 @@ const Lobby = () => {
     return () => {
       socket.off("joined-lobby", onLobbyJoin);
       socket.off("joined-team", onTeamJoin);
-      socket.off("global-message-received", onGlobalMessage);
-      socket.off("team-message-received", onTeamMessage);
       socket.off("team-name-edited", onTeamNameEdited);
       socket.off("player-disconnect", onPlayerDisconnect);
       socket.off("teams-reset", onTeamsReset);
       socket.off("teams-randomized", onTeamsRandomized);
     };
-  }, [currentTab, numOfUnreadMessages]);
+  }, [messages]);
 
   return (
     <>
@@ -181,61 +160,7 @@ const Lobby = () => {
                 />
               </Background>
               <Background className="col-span-2 col-start-5 row-span-1 row-start-2">
-                <Tabs.Root
-                  className="flex h-full flex-col p-2"
-                  value={currentTab}
-                  onValueChange={(val) => {
-                    setCurrentTab(val);
-                    setNumOfUnreadMessages(0);
-                  }}
-                >
-                  <Tabs.List className="flex pb-2">
-                    <Tabs.Trigger
-                      className="relative flex w-full items-center justify-center border-b-purple-600 pb-1 text-lg font-bold data-[state=active]:border-b"
-                      value="global"
-                    >
-                      <div className="relative">
-                        Global
-                        {numOfUnreadMessages > 0 && currentTab !== "global" && (
-                          <div className="absolute -top-2 -right-7 inline-flex h-5 w-5 items-center justify-center rounded-full bg-red-500 text-sm font-bold text-white">
-                            {numOfUnreadMessages}
-                          </div>
-                        )}
-                      </div>
-                    </Tabs.Trigger>
-                    {user.team && (
-                      <Tabs.Trigger
-                        className="flex w-full items-center justify-center border-b-purple-600 pb-1 text-lg font-bold data-[state=active]:border-b"
-                        value="team"
-                      >
-                        <div className="relative">
-                          Team
-                          {numOfUnreadMessages > 0 && currentTab !== "team" && (
-                            <div className="absolute -top-2 -right-7 inline-flex h-6 w-6 items-center justify-center rounded-full bg-red-500 text-xs font-bold text-white">
-                              {numOfUnreadMessages}
-                            </div>
-                          )}
-                        </div>
-                      </Tabs.Trigger>
-                    )}
-                  </Tabs.List>
-
-                  <Tabs.Content value="global" className="h-full overflow-clip">
-                    <Chat messages={messages} userName={user.name} />
-                  </Tabs.Content>
-
-                  <Tabs.Content value="team" className="h-full overflow-clip">
-                    <Chat
-                      messages={teamMessages}
-                      userName={user.name}
-                      roomId={
-                        router.query.id && user.team
-                          ? router.query.id + user.team
-                          : ""
-                      }
-                    />
-                  </Tabs.Content>
-                </Tabs.Root>
+                <ChatCard user={user} />
               </Background>
               <div className="col-start-1 col-end-7">
                 <div className="flex h-full gap-4">
@@ -247,7 +172,6 @@ const Lobby = () => {
                       isQuizMaster={isQuizMaster}
                       name={team.name}
                       players={team.players}
-                      setTeamMessages={setTeamMessages}
                       setUser={setUser}
                       user={user}
                     />
